% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp.dm-package.R
\docType{package}
\name{amp.dm-package}
\alias{amp.dm}
\alias{amp.dm-package}
\title{amp.dm: Data Management Tools for Pharmacometrics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools and functions to efficiently create datasets used in pharmacometric analysis. Additional functionality is added to create documentation and prepare files for submission and quality control purposes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://leidenadvancedpkpd.github.io/amp.dm/}
  \item \url{https://github.com/LeidenAdvancedPKPD/amp.dm}
  \item Report bugs at \url{https://github.com/LeidenAdvancedPKPD/amp.dm/issues}
}

}
\author{
\strong{Maintainer}: Richard Hooijmaijers \email{richardhooijmaijers@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item LAPP Consultants \email{info@lapp.nl} [funder, copyright holder]
}

}
\keyword{internal}
