% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_add.r
\name{attr_add}
\alias{attr_add}
\title{Add attributes from a list to a dataframe}
\usage{
attr_add(data, attrl, attrib = c("label", "format", "remark"), verbose = TRUE)
}
\arguments{
\item{data}{the data frame for which the attributes should be set}

\item{attrl}{named list with the attributes for the dataset (see details)}

\item{attrib}{a vector of attributes that should be set for data (currently 'label', 'format' and 'remark' are applicable)}

\item{verbose}{a logical indicating if datachecks should be printed to console}
}
\value{
dataframe with the attributes added
}
\description{
This function adds data attributes available in a list to a data frame. Additional checks
are done to verify if the attributes are in a valid and use-able format
}
\details{
This function adds attributes available in a list to a data frame. The structure of this list should be available
in a specific format. The names items in the list are aligned with the variables in the data frame. For each item,
the content of the 'label', 'format' and 'remark' elements will be added as attributes to the dataset. For an example
of the format of list see for instance \link{attr_xls}.
}
\examples{

data(esoph)
attr(esoph$ncontrols,'label')   <- 'number of controls'
tmp <- subset(esoph,,-agegp)
str(tmp)
}
\seealso{
\link{attr_xls}
}
\author{
Richard Hooijmaijers
}
\keyword{attribute}
