% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_extract.r
\name{attr_extract}
\alias{attr_extract}
\title{Reads in attributes from an external excel file}
\usage{
attr_extract(dfrm)
}
\arguments{
\item{dfrm}{data frame containing the attributes}
}
\value{
named list with the attributes
}
\description{
This function extracts attributes available in a data frame and creates a structured list
}
\examples{
attrl  <- attr_xls(system.file("example/Attr.Template.xlsx",package = "amp.dm"))
nm     <- read.csv(system.file("example/NM.theoph.V1.csv",package = "amp.dm"))
nmf    <- attr_add(nm, attrl, verbose = FALSE)
attrl2 <- attr_extract(nmf)
all.equal(attrl,attrl2)
}
\author{
Richard Hooijmaijers
}
\keyword{documentation}
