% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_factor.r
\name{attr_factor}
\alias{attr_factor}
\title{Create factors for variables within a data frame using the format attribute}
\usage{
attr_factor(data, verbose = TRUE)
}
\arguments{
\item{data}{the data.frame for which factors should be created}

\item{verbose}{a logical indicating if datachecks should be printed to console}
}
\value{
data frame with the formats assigned
}
\description{
This function searches for the 'format' attribute within a data frame, if found it applies the format to
that variable. The resulting variable will be a factor useful for plotting and reporting
}
\details{
In order to make this function work the 'format' attribute should be present and should be available
as a named vector (e.g. \code{attr(data$GENDER,'format')   <- c('0' = 'Male', '1' = 'Female')}). If the
attribute is found it overwrites the variable with the format applied to it. Be aware that the original
levels defined in the format could be lost. The \code{factor} function in R always sets the levels from 1 to n.
This means that if a format does not start with 1 or skips categories the underlying levels may differ.
}
\examples{

tmp <- data.frame(GENDER=rep(c(0,1),4),RESULT=rnorm(8))
attr(tmp$GENDER,'format')   <- c('0' = 'Male', '1' = 'Female')
tmp <- attr_factor(tmp)
str(tmp$GENDER)
}
\seealso{
\link{factor}, \link{attr_add}, \link{attr_xls}
}
\author{
Richard Hooijmaijers
}
\keyword{attribute}
