% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_xls.r
\name{attr_xls}
\alias{attr_xls}
\title{Reads in attributes from an external excel file}
\usage{
attr_xls(xls, sepfor = "\\n", nosort = FALSE)
}
\arguments{
\item{xls}{character with the name of the excel file containing the attributes}

\item{sepfor}{character of length 1 indicating what the seperator for formats should be}

\item{nosort}{logical indicating if sorting of variables should be omitted (otherwise sorting of no. column in excel file is applied)}
}
\value{
named list with the attributes
}
\description{
This function reads in attributes available in an excel file and creates a strcutured list
}
\examples{
\dontrun{
  xmpl  <- system.file("example/Attr.Template.xlsx",package = "amp.dm")
  attr_xls(xmpl)
}
}
\author{
Richard Hooijmaijers
}
\keyword{documentation}
