% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_arg}
\alias{check_arg}
\title{Function that checks arguments of parent function}
\usage{
check_arg(..., data, type = 1)
}
\arguments{
\item{...}{arguments from parent function to check}

\item{data}{data frame needed for type 3}

\item{type}{numeric with the type to check (see details)}
}
\value{
named character string with the result
}
\description{
Function checks the values and type of arguments of a parent function
}
\details{
The function can perform various checks. Either the type of argument is returned (1)
wether the argument is NULL (2), what the value of the argument is as character string (3)
or if the arguments are available in the data (4)
}
\examples{
\dontrun{
  test1 <- function(aa,bb)  check_arg(aa,bb,type = 1)
  test1(aa="a",bb=b)
  test2 <- function(cc = NULL)  check_arg(cc,type = 2)
  test2()
  test3 <- function(dd,data)  check_arg(dd,data=data,type = 4)
  test3(dd="nam",data=data.frame(nam=1))
}
}
\author{
Richard Hooijmaijers
}
\keyword{helper}
