% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cat.r
\name{check_cat}
\alias{check_cat}
\title{Create an overview of the available categories}
\usage{
check_cat(x, missing = c(-999, NA), detail = 5, threshold = c(NA, NA))
}
\arguments{
\item{x}{numeric vector with the categories}

\item{missing}{vector with the values that present missing information}

\item{detail}{numeric witht he level of detail to print (see below for details)}

\item{threshold}{numeric vector with the threshold numbers and proportions (see details)}
}
\value{
Nothing is returned information is printed on screen
}
\description{
This function reports information for the categories, mainly the frequencies, proportions
and missing values
}
\details{
The detail argument can be used to print certain information:
\itemize{
\item 5: All possible information is printed
\item 4: Only the table with frequencies and proportions
\item 3: Only information regarding missing data
\item 2: Only a warning in case number of missing is above threshold (see below)
\item 1: A named vector with the available categories that can be used in \link{num_lump}
The threshold presents the absolute number (first number) and proportions (second number) to check.
If either one of these numbers is above the threshold for missing values, a warning is given.
This can be convenient to decide whether or not a category should be used during analyses.
}
}
\examples{

data1   <- data.frame(cat1 = c(rep(1:5,10),-999), 
                      cat2 = c(rep(letters[1:5],10),-999))
check_cat(data1$cat1)
check_cat(data1$cat2, detail=1)                       
check_cat(data1$cat1,detail=2,threshold = c(NA,0.1))
}
\author{
Richard Hooijmaijers
}
\keyword{documentation}
