% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nmdata.r
\name{check_nmdata}
\alias{check_nmdata}
\title{Checks nonmem dataset for common errors/mistakes}
\usage{
check_nmdata(
  x,
  type = 1,
  ret = "tbl",
  capt = NULL,
  align = NULL,
  size = "\\\\footnotesize",
  ...
)
}
\arguments{
\item{x}{either a path to a CSV file or a data frame  with the NONMEM data that should be checked}

\item{type}{integer with the type of checks. Currently 1 can be used for checks that should all pass for a valid analysis
and 2 for checks that trigger further investigation}

\item{ret}{a character vector to define what kind of output should be returned (either "dfrm", "tbl", "file")}

\item{capt}{character with the caption of the table (not used in case data frame is returned)}

\item{align}{alignment of the table passed to \link{general_tbl} (not used in case data frame is returned)}

\item{size}{character with font size as for the table \link{general_tbl}}

\item{...}{additional arguments passed to \link{general_tbl}}
}
\value{
the checks are either printed, returned as dataframe or placed in a PDF
}
\description{
This function checks if there are any common errors or mistakes within a NONMEM dataset,
and reports the results back to console, table or dataframe
}
\examples{

\dontrun{
  check_nmdata("nonmemData.csv")
  check_nmdata(dataframe)
}
}
\author{
Richard Hooijmaijers
}
\keyword{manip}
