% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cmnt}
\alias{cmnt}
\title{Add comment to environment to present in documention}
\usage{
cmnt(string = "", bold = FALSE, verbose = TRUE)
}
\arguments{
\item{string}{character of length one with the comment to add}

\item{bold}{logical indicating if the string should be printed in bold to emphasize importance}

\item{verbose}{logical indicating if the comment should be printed when function is called}
}
\value{
no return value, called for side effects
}
\description{
Adds a comment regarding assumptions and special attention into package environment,
which can be used in code chunks and easily printed after a code chunk
}
\examples{
  cmnt("Exclude time points > 12h")
  cmnt("Subject 6 deviates and is excluded in the analysis", TRUE)
  # Markdown syntax can be used for comments:
  cmnt("We can use **bold** and *italic* or `code`")
  # we can print the contents of the comments with
  get_log()$cmnt_nfo
}
\author{
Richard Hooijmaijers
}
\keyword{documentation}
