% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cmnt_print}
\alias{cmnt_print}
\title{Function that prints the comments given by \link{cmnt}}
\usage{
cmnt_print(clean = TRUE)
}
\arguments{
\item{clean}{logical indicating if the comments should be deleted after printing (see details)}
}
\value{
character string with the comments
}
\description{
Prints the results in markdown format to be used directly in inline coding
}
\details{
The function returns a text string with the comments given up to the point
it was called. When clean is set to TRUE (default), the content of the comment
dataset is cleaned to overcome repetition of comments each time it is called
}
\examples{
  cmnt("Comment to print")
  cmnt_print()
}
\author{
Richard Hooijmaijers
}
\keyword{documentation}
