% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents_df.r
\name{contents_df}
\alias{contents_df}
\title{Create information for multiple data frames}
\usage{
contents_df(
  dfv,
  subject = NULL,
  ret = "tbl",
  capt = "Information multiple data frames",
  align = "lllp{8cm}",
  ...
)
}
\arguments{
\item{dfv}{a character vector with data frame(s) in global environment for which the overview should be created}

\item{subject}{character string that identifies the subject variable within the data frame}

\item{ret}{a character vector to define what kind of output should be returned (either "dfrm", "tbl", "file")}

\item{capt}{character with the caption of the table (not used in case data frame is returned)}

\item{align}{alignment of the table passed to \link{general_tbl} (not used in case data frame is returned)}

\item{...}{additional arguments passed to \link{general_tbl}}
}
\value{
a data frame, code for table or nothing in case a PDF file is created
}
\description{
This function creates a latex table or data frame with the number of records, subjects and variables
of one or multiple data frames.
}
\details{
This function can be used to create a table with the most important information of a data
frame for documentation. The function will list the the number of records, subjects and variables
of each data frame within dfv. This function is especially usable to indicate the differences between
similar data frames or an overview of all data frames within a working environment
}
\examples{

Theoph1 <-  subset(Theoph,Subject!=1)
Theoph2 <-  subset(Theoph,Subject!=2)
contents_df(c('Theoph1','Theoph2'),subject='Subject',ret='dfrm')
}
\author{
Richard Hooijmaijers
}
\keyword{documentation}
