% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_tbl.r
\name{define_tbl}
\alias{define_tbl}
\title{Create define PDF file for submission of pharmacometric data files}
\usage{
define_tbl(
  attr = NULL,
  ret = "dfrm",
  capt = "Dataset define form",
  align = "lllp{10cm}",
  outnm = NULL,
  orientation = "portrait",
  ...
)
}
\arguments{
\item{attr}{list with datasets attributes}

\item{ret}{a character vector to define what kind of output should be returned (either "dfrm", "tbl", "file")}

\item{capt}{character with the caption of the table}

\item{align}{alignment of the table passed to \link{general_tbl}}

\item{outnm}{character with the name of the tex file to generate and compile (e.g. define.tex)}

\item{orientation}{character the page orientation in case a file is to be returned (can be either 'portrait' or 'landscape')}

\item{...}{additional arguments passed to \link{general_tbl}}
}
\value{
a data frame, code for table or nothing in case a PDF file is created
}
\description{
This function creates the define.pdf file necessary for esubmission.
}
\examples{

\dontrun{definePDF(attrl,outnm='define.tex')}
}
\author{
Richard Hooijmaijers
}
