% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_addl_ii.r
\name{expand_addl_ii}
\alias{expand_addl_ii}
\title{Expand rows in case ADDL and II variables are present}
\usage{
expand_addl_ii(data, evid = NULL, del_iiaddl = TRUE)
}
\arguments{
\item{data}{data frame to perform the expantion on}

\item{evid}{<\code{\link[rlang:args_data_masking]{data-masking}}> identifies the event ID (EVID) within the data frame.
This is used to distinguish observations from dosing records, e.g. 0 for observations}

\item{del_iiaddl}{logical identifying if the ADDL and II variables can be deleted from output}
}
\value{
a data frame with expanded dose records
}
\description{
This function expands ADDL and II records. This is done by placing each ADDL record
on a separate line. This is convenient in case of individual dose calculations
}
\details{
The function expects that certain variables are present in the data (at least ID, TIME, ADDL and II)
}
\examples{

dfrm <- data.frame(ID=c(1,1), TIME=c(0,12),II=c(12,0),ADDL=c(5,0),AMT=c(10,0),EVID=c(1,0))
expand_addl_ii(dfrm,evid="EVID")
}
\author{
Richard Hooijmaijers
}
\keyword{manipulation}
