% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dates.r
\name{fill_dates}
\alias{fill_dates}
\title{Fills down dates within a data frame that include a start and end date}
\usage{
fill_dates(data, start, end, tau = 1, repdat = 1)
}
\arguments{
\item{data}{data frame for which the dates should be filled down}

\item{start}{<\code{\link[rlang:args_data_masking]{data-masking}}> that includes the start date (as date format)}

\item{end}{<\code{\link[rlang:args_data_masking]{data-masking}}> that includes the end date (as date format)}

\item{tau}{integer with the tau in days (e.g. 2 for dosing every other day)}

\item{repdat}{integer with repeats per day (e.g. 2 in case of twice daily dosing)}
}
\value{
a data frame with filled out dates
}
\description{
This function can be used in case a start and end date is known for dosing. This function fills down
the dates so each date between start and end is placed on a separate row. Subsequently the dataset
can be used to merge with available date information and impute missing dates.
}
\examples{

dfrm <- data.frame(ID=1:2,first=as.Date(c("2016-10-01","2016-12-01"), "\%Y-\%m-\%d"),
                   last=as.Date(c("2016-10-03","2016-12-02"), "\%Y-\%m-\%d"))
fill_dates(dfrm, first, last)
fill_dates(dfrm, first, last, 2, 3)
}
\author{
Richard Hooijmaijers
}
\keyword{manipulation}
