% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterr.r
\name{filterr}
\alias{filterr}
\title{Filter data with logging of results}
\usage{
filterr(.data, ..., comment = "")
}
\arguments{
\item{.data}{the data frame for which the filter should be created}

\item{...}{arguments passed to \link[dplyr:filter]{dplyr::filter}}

\item{comment}{character with the reason of filtering used in log file}
}
\value{
a filtered data frame
}
\description{
This function is a wrapper around \link[dplyr:filter]{dplyr::filter}. Additional actions are performed on the background to
log the information of the filter action, and info regarding the step is printed.
}
\details{
The function can be used to keep track of records that are omitted in the data management process.
In general one would like to keep all records from the source database (and use flags instead to exclude data) but
in cases where this is not possible it is important to know what records are omitted and for which reason.
Every time the function is used it creates a records in in a log file which can be used in the documentation.
}
\examples{

# For full trace-ability of source data, no pipes or 
# base R pipes are preferred 
dat1 <- filterr(Theoph,Subject==1)
dat2 <- Theoph |> filterr(Subject==2)
dat3 <- Theoph \%>\% filterr(Subject==3)
# Show what is being logged
get_log()$filterr_nfo
}
\seealso{
\link[dplyr:filter]{dplyr::filter}
}
\author{
Richard Hooijmaijers
}
\keyword{manipulation}
