% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_outliers.r
\name{flag_outliers}
\alias{flag_outliers}
\title{Creates a flag for outlying values}
\usage{
flag_outliers(var, type = "boxstat")
}
\arguments{
\item{var}{numeric vector that should be checked for outliers}

\item{type}{character with the type of test to perform for outliers (currently only the "boxstats" is available that uses the \link{boxplot} method)}
}
\value{
a numeric vector the same length as \code{var} with either 0 (no outlier) or 1 (outlier)
}
\description{
This function creates a flag identifying the outliers in a vector
}
\examples{

dfrm <- data.frame(a = 1:10, b = c(1:9,50))
flag_outliers(dfrm$a)
flag_outliers(dfrm$b)
}
\author{
Richard Hooijmaijers
}
\keyword{manipulation}
