% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tbl.r
\name{general_tbl}
\alias{general_tbl}
\title{General table wrapper for documentation functions}
\usage{
general_tbl(
  data,
  ret = "tbl",
  capt = "General table",
  align = NULL,
  outnm = NULL,
  tabenv = "longtable",
  float = FALSE,
  hlineafter = NULL,
  addtorow = list(pos = list(-1, 0), command = c("\\\\toprule ",
    "\\\\midrule\\\\endhead ")),
  ...
)
}
\arguments{
\item{data}{data frame for which the table should be created}

\item{ret}{a character vector to define what kind of output should be returned (either "dfrm", "tbl", "file")}

\item{capt}{character with the caption of the table}

\item{align}{alignment of the table passed to \link[R3port:ltx_list]{R3port::ltx_list} or \link[xtable:print.xtable]{xtable::print.xtable} (see details below)}

\item{outnm}{character with the name of the tex file to generate and compile, e.g. "define.tex" (only applicable in case file is returned)}

\item{tabenv}{character with the tabular environment passed to \link[xtable:print.xtable]{xtable::print.xtable}}

\item{float}{logical defining float passed to \link[xtable:print.xtable]{xtable::print.xtable}}

\item{hlineafter}{a vector indicating the rows after which a horizontal line should appear passed to \link[xtable:print.xtable]{xtable::print.xtable}}

\item{addtorow}{A list containing the position of rows and commands passed to \link[xtable:print.xtable]{xtable::print.xtable}}

\item{...}{additional arguments passed to either \link[R3port:ltx_list]{R3port::ltx_list} or \link[xtable:print.xtable]{xtable::print.xtable} depending on what is returned}
}
\value{
a data frame, code for table or nothing in case a PDF file is created
}
\description{
This function creates a latex table
}
\details{
This function is a general function to create a xtable suitable for documentation, or directly creates an compiles a
latex file using the \code{R3port} package. The align argument can be used to change the column widths of a table to be able to fit a table
on a page (e.g. in latex wide tables can fall off a page). The way this should be used is to provide a vector of one with
a specification for each column. For example \verb{lp\{3cm\}r} can be used for left align first column, second column of 3 cm
and third column right aligned. In case align is set to NULL a default alignment is used
}
\examples{

general_tbl(head(Theoph))
}
\author{
Richard Hooijmaijers
}
\keyword{internal}
