% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_script.r
\name{get_script}
\alias{get_script}
\title{Get the current script name (either interactive Rstudio, markdown or batch script)}
\usage{
get_script(base = TRUE, noext = TRUE)
}
\arguments{
\item{base}{logical indicating if the basename should be returned (without path)}

\item{noext}{logical indicating if the file extension should be omitted}
}
\value{
character with the current script name
}
\description{
Get the current script name (either interactive Rstudio, markdown or batch script)
}
\author{
Richard Hooijmaijers
}
