% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_covar.r
\name{impute_covar}
\alias{impute_covar}
\title{Imputeute missing covariates}
\usage{
impute_covar(var, uniques = NULL, type = "median", verbose = FALSE)
}
\arguments{
\item{var}{vector with the items that should be imputed}

\item{uniques}{vector that defines unique records to enable calculation of stats on non duplicate values}

\item{type}{character of length one defining the type of statistics to perform for imputation (see details)}

\item{vernose}{logical indicating if additional information should be given}
}
\value{
a vector where missing values are imputed
}
\description{
The function will impute all NA values with either a given statistic (e.g. median) or with the largest group
}
\details{
The function can be used to impute continuous or categorical covariates. In case continuous covariates the type
argument should be a statistic like median or mean. In case a categorical covariate is used, the type should be set
to 'largest' in which case the category that occurs most is used. In case multiple values occur most, the last encountered
is used.
}
\examples{
dfrm  - data.frame(num1 = c(NA,110))
impute_covar(dfrm$num1,type=median)
}
\author{
Richard Hooijmaijers
}
\keyword{manipulation}
