% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_dose.r
\name{impute_dose}
\alias{impute_dose}
\title{Imputes dose records using ADDL and II by looking forward and backwards}
\usage{
impute_dose(data, id, datetime, amt = "AMT", ii = "II", thr = 50, back = TRUE)
}
\arguments{
\item{data}{data frame to perform the calculations on}

\item{id}{character identifying the id variable within the data frame}

\item{datetime}{character identifying the date/time variable (POSIXct) within the data frame}

\item{amt}{character identifying the AMT (amount) variable within the data frame}

\item{ii}{character identifying the II (Interdose Interval)  variable within the data frame}

\item{thr}{numeric indicating the threshold percentage for imputation (see details)}

\item{back}{logical indicating if the time for imputed doses should be taken from the previous record (TRUE) or the next (FALSE)}
}
\value{
a data frame with imputed dose records
}
\description{
This function imputes dose records by looking at the missing doses between available records
based on a given II value
}
\details{
The function fills in the doses by looking at the time difference and  II between all records.
Be aware that this can result in unexpected results in a few cases, so results should always be handled with care.
The function will determine if a dose should be imputed based on the 'thr' value. For each dose, the function detemines the percentage
difference from the previous dose based on the II value. In case this the expected difference is above the threshold value imputation will be done.
}
\examples{

dfrm <- data.frame(id=c(1,1), dt=c(Sys.time(),Sys.time()+ 864120),
                   II=c(24,24),AMT=c(10,10))
impute_dose(dfrm,"id","dt")
}
\author{
Richard Hooijmaijers
}
\keyword{manipulation}
