% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_df.r
\name{log_df}
\alias{log_df}
\title{Create information for all functions that log actions}
\usage{
log_df(
  log,
  type,
  coding = FALSE,
  ret = "dfrm",
  capt = NULL,
  align = NULL,
  size = "\\\\footnotesize",
  ...
)
}
\arguments{
\item{log}{list with logged information typically obtained with \link{get_log}}

\item{type}{character with the type of info that should be taken from log (either "filterr_nfo","joinr_nfo" or "read_nfo")}

\item{coding}{logical indicating if the coding (within \code{filterr}) should be displayed}

\item{ret}{a character vector to define what kind of output should be returned (either "dfrm", "tbl", "file")}

\item{capt}{character with the caption of the table (not used in case data frame is returned)}

\item{align}{alignment of the table passed to \link{general_tbl} (not used in case data frame is returned)}

\item{size}{character with font size as for the table \link{general_tbl}}

\item{...}{additional arguments passed to \link{general_tbl}}
}
\value{
function creates a PDF file or returns a data frame
}
\description{
This function creates a table including information on functions that log information such as
\code{filterr}, \code{left_joinr} and \code{read_data}
}
\details{
This function generates information for function that logs information. It attempts to find a good alignment
and caption, mainly for outputting to a table. It is possible to set your own captions and alignment, take into
account that the alignment differs per type and in case the coding argument is changed.
}
\examples{

dat1 <- filterr(Theoph,Subject==1)
dat2 <- Theoph |> filterr(Subject==2)
log_df(get_log(), "filterr_nfo")
}
\author{
Richard Hooijmaijers
}
\keyword{documentation}
