% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_lump.r
\name{num_lump}
\alias{num_lump}
\title{Perform lumping of numerical values}
\usage{
num_lump(x, lumpcat = 99, uniques = NULL, prop = NULL, min = NULL, ...)
}
\arguments{
\item{x}{numeric vector with the items that should be lumped}

\item{lumpcat}{the category in which the lumped levels should be added (see details)}

\item{uniques}{vector that defines unique records to enable lumping on non duplicate values}

\item{prop}{numeric with the threshold proportions for lumping}

\item{min}{numeric with the min number of times a level should appear to not lump}

\item{...}{additional arguments passed to \link[forcats:fct_lump]{forcats::fct_lump_min} and/or \link[forcats:fct_lump]{forcats::fct_lump_prop}}
}
\value{
vector with the lumping applied
}
\description{
This function is mainly a wrapper for \link[forcats:fct_lump]{forcats::fct_lump} but applied on numeric variables. Furthermore
there is the option to use uniques to determine small categories for instance on individual level
}
\details{
The argument lumpcat is the level in which lumped values should appear and can be one of the following:
\itemize{
\item numeric with the category number to set the levels to
\item character specifying "largest" to select the largest category (selected before lumping)
\item named vector to set the 'algorithm' for instance:
c('5'='3', '4'='6') to set category 5 to 3 and 4 to 6 when these categories need lumping
}
}
\examples{

dfrm <- data.frame(id = 1:30, cat = c(rep(1,8),rep(2,13), rep(3,4),rep(4,5)))
num_lump(x=dfrm$cat, lumpcat=99, prop=0.15)
}
\author{
Richard Hooijmaijers
}
\keyword{manipulation}
