% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_data.r
\name{output_data}
\alias{output_data}
\title{export R data for NONMEM modeling}
\usage{
output_data(
  x,
  csv = NULL,
  xpt = NULL,
  attr = NULL,
  verbose = TRUE,
  maxdig = 6,
  tonum = TRUE,
  firstesc = NULL,
  readonly = FALSE,
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{x}{data frame to be exported.}

\item{csv}{character with the name of the csv file to generate}

\item{xpt}{character with the name of the xpt file to generate}

\item{attr}{character with the name of the rds file to generate}

\item{verbose}{logical indicating if additional information should be written to the console}

\item{maxdig}{numeric with the maximum number of decimals for numeric variables to be in output (see details)}

\item{tonum}{logical indicating if all variables should be made numeric (standard for NONMEM input files)}

\item{firstesc}{character with escape character for first variable, used to exclude row in NONMEM}

\item{readonly}{logical indicating if the output csv file should be made readonly}

\item{overwrite}{logical indicating if (all) output should be overwritten or not}

\item{...}{Arguments passed to \code{\link{write.csv}}}
}
\value{
a data frame is written to disk
}
\description{
This function exports data for NONMEM modeling analyses including
options that are frequently necessary to adapt.
}
\details{
In case tonum is \code{TRUE}, all variables will be made numeric and \code{Inf} values will be
set to \code{NA} (all \code{NA} values will be set to a dot). The rounding set in \code{maxdig}
will only be done in case tonum is set to TRUE. For xpt files, the name of the object to export is used as the name
of the dataset inside the xpt file (e.g. \code{output_data(dfrm,xpt='dataset.xpt')} will result in an xpt file named
'dataset.xpt' with one dataset named 'dfrm').
}
\examples{
\dontrun{
  data(Theoph)
  out_file <- tempfile(fileext = ".csv")
  output_data(Theoph, csv = out_file, tonum = FALSE)
}
}
\seealso{
\code{\link{write.csv}}
}
\author{
Richard Hooijmaijers
}
\keyword{IO}
\keyword{file}
