% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vars.r
\name{plot_vars}
\alias{plot_vars}
\title{Creates different kind of plots for variables within a dataset using ggplot2}
\usage{
plot_vars(dfrm, vars = names(dfrm), ppp = 16, ...)
}
\arguments{
\item{dfrm}{data frame that should be plotted}

\item{vars}{character vector with the variables for which plots should be generated}

\item{ppp}{number plots per page}

\item{...}{additional arguments passed to \code{\link[patchwork:wrap_plots]{patchwork::wrap_plots()}} (e.g. ncol/nrow)}
}
\value{
a ggplot/patchwork object with all variables plotted visualized
}
\description{
This function creates histograms for numeric values and barcharts for character or factor variables
In case there are more then 10 unique values it will list the first 10 unique values in a 'text' plot
}
\examples{

plot_vars(Theoph)
}
\author{
Richard Hooijmaijers
}
\keyword{misc}
