% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_tbl.r
\name{session_tbl}
\alias{session_tbl}
\title{Create information for R session}
\usage{
session_tbl(ret = "tbl", capt = "Session info", align = "lp{8cm}", ...)
}
\arguments{
\item{ret}{a character vector to define what kind of output should be returned (either "dfrm", "tbl", "file")}

\item{capt}{character with the caption of the table (not used in case data frame is returned)}

\item{align}{alignment of the table passed to \link{general_tbl} (not used in case data frame is returned)}

\item{...}{additional arguments passed to \link{general_tbl}}
}
\value{
a data frame, code for table or nothing in case a PDF file is created
}
\description{
This function creates a latex table or data frame with information from the R session
(sessionInfo() and Sys.info())
}
\details{
This function can be used to create a table with the most important information of a R session,
the user that is running the R session and the current date/time
}
\examples{

session_tbl()
}
\author{
Richard Hooijmaijers
}
\keyword{documentation}
