% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{srce}
\alias{srce}
\title{Add source information to environment to present in documention}
\usage{
srce(var, source, type = "c")
}
\arguments{
\item{var}{unquoted string with the variable for which the source should be defined}

\item{source}{unquoted strings with the source(s) used for var (see example)}

\item{type}{character with the type of variable can be either 'c' (copied) or 'd' (derived)}
}
\value{
no return value, called for side effects
}
\description{
Adds the source of variables into package environment,
which can be used in code chunks at the applicable locations and easily added to
documention afterwards
}
\examples{
  # variable AMT copied from Dose variable in Theoph data frame
  srce(AMT,Theoph.Dose)
  # variable BMI derived from WEIGHT variable in wt data frame
  # and HEIGHT variable in ht data frame
  srce(BMI,c(wt.WEIGHT,ht.HEIGHT),'d')
  get_log()$srce_nfo
}
\author{
Richard Hooijmaijers
}
\keyword{documentation}
