% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_df.r
\name{stats_df}
\alias{stats_df}
\title{Calculate basic statistics on data frame}
\usage{
stats_df(
  data,
  missingval = -999,
  ret = "tbl",
  capt = "Statistics data frame",
  align = "p{2cm}p{1cm}p{1cm}p{4cm}p{1.7cm}p{1.7cm}p{0.8cm}p{1.3cm}",
  size = "\\\\footnotesize",
  ...
)
}
\arguments{
\item{data}{a data frame for which the overview should be created}

\item{missingval}{numeric with the value that indicates missing values, if NULL no missings will be recorded}

\item{ret}{a character vector to define what kind of output should be returned (either "dfrm", "tbl", "file")}

\item{capt}{character with the caption of the table (not used in case data frame is returned)}

\item{align}{alignment of the table passed to \link{general_tbl} (not used in case data frame is returned)}

\item{size}{character with font size as for the table \link{general_tbl}}

\item{...}{additional arguments passed to \link{general_tbl}}
}
\value{
either tex code for table a data frame
}
\description{
This function creates a latex table or data frame with the basic statistics of a
data frame.
}
\details{
This function can be used to create a table with basic statistics of a data
frame. The function will list the min, max, number of NA/missing values, number of unique categories and
type of variable of all data items within a data frame. In case a data item has less than 10 unique categories, it
will list the unique values.
The main reason to use this function is to create a structured table with
statistics of a data frame to be included in documentation.
}
\examples{

stats_df(Theoph)
}
\author{
Richard Hooijmaijers
}
\keyword{documentation}
