% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{typearg}
\alias{typearg}
\title{Function that takes the formals of a function and return the type}
\usage{
typearg(frml)
}
\arguments{
\item{frml}{a list typically obtained by the \link{formals} function}
}
\value{
named vector with the type of values for the function arguments
}
\description{
function get the \link{typeof} of a function arguments but also reports na, missing, null and empty
}
\examples{
  typearg(formals(lm))
}
\author{
Richard Hooijmaijers
}
\keyword{documentation}
