% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_height.r
\name{weight_height}
\alias{weight_height}
\title{Calculates different weight variables}
\usage{
weight_height(wt = NULL, ht = NULL, sex = NULL, bmi = NULL, type = "bmi")
}
\arguments{
\item{wt}{vector with weight values, in either kg or lb depending on the type (see details)}

\item{ht}{vector with height values in cm (see details)}

\item{sex}{vector with SEX values (Where female is defined as a value of 1)}

\item{bmi}{vector with BMI values (see details)}

\item{type}{character with the type to be used for the calculations (see details)}
}
\value{
a vector with calculated values
}
\description{
This function calculates different variables based on weight and height and
conversion from or to kilograms
}
\details{
Currently the following types are defined within the function:
\subsection{Convert units}{
\itemize{
\item "kg-lb" : Convert units from kg to lb using the formula \deqn{\textrm{Weight (kg)} = \textrm{Weight (lb)} \cdot 2.20462262}
\item "lb-kg" : Convert units from lb to kg using the formula \deqn{\textrm{Weight (lb)} = \frac{\textrm{Weight (kg)}}{2.20462262}}
}
}

\subsection{Body mass index}{
\itemize{
\item "bmi" :  Calculates body mass index (BMI) using the standard formula (\href{https://www.cambridge.org/core/books/treatise-on-man-and-the-development-of-his-faculties/AB13A647A6C8727C06AE5399D7422887}{Quetelet1842},
\deqn{\textrm{BMI} = \frac{\textrm{Weight (kg)}}{\textrm{Height (m)}^{2}}}
}
}

\subsection{Body Surface Area}{
\itemize{
\item "bsa": Body Surface Area, according to \href{https://www.researchgate.net/publication/17656913_Estimation_of_Human_Body_Surface_Area_from_Height_and_Weight}{Gehan and Georg},
\deqn{\textrm{BSA} = \exp{-3.751} \cdot \textrm{Height (cm)}^{0.422} \cdot \textrm{Weight (kg)}^{0.515}}
\item "bsa2": Body Surface Area, according to \href{https://pubmed.ncbi.nlm.nih.gov/2520314/}{DuBois and DuBois},
\deqn{\textrm{BSA} = 0.007184 \cdot \textrm{Height (cm)}^{0.725} \cdot \textrm{Weight (kg)}^{0.425}}
\item "bsam": Body Surface Area, according to \href{https://pubmed.ncbi.nlm.nih.gov/3657876/}{Mosteller},
\deqn{\textrm{BSA} = \sqrt{\frac{\textrm{Weight (kg)} \cdot \textrm{Height (cm)}}{3600}}}
\item "bsah": Body Surface Area, according to \href{https://pubmed.ncbi.nlm.nih.gov/650346/}{Haycock},
\deqn{\textrm{BSA} = 0.024265 \cdot \textrm{Height (cm)}^{0.3964} \cdot \textrm{Weight (kg)}^{0.5378}}
\item "bsal": Body Surface Area in normal-weight and obese adults up to 250 kg, according to \href{https://pubmed.ncbi.nlm.nih.gov/11500314/}{Livingston},
\deqn{\textrm{BSA} = 0.1173 \cdot \textrm{Weight (kg)}^{0.6466}}
}
}

\subsection{Fat free mass}{
\itemize{
\item "ffmj": Fat free mass, according to \href{https://pubmed.ncbi.nlm.nih.gov/16176118/}{Janmahasatian}:
\deqn{\textrm{FFM}=\frac{9270 \cdot \textrm{Weight (kg)}}{k +\left(l \cdot  \textrm{BMI}\right)}}, where \eqn{k} is 6680 for males and 8780 for females and \eqn{l} is 216 for males and 244 for females.
\item "ffms": Fat free mass in Indian patients, according to \href{https://pubmed.ncbi.nlm.nih.gov/32201910/}{Sinha}:
\deqn{\textrm{FFM}=\frac{9270 \cdot \textrm{Weight (kg)}}{k \cdot  l \cdot \textrm{BMI}^{0.28}}}, where \eqn{k} is 6680 for males and 8780 for females and \eqn{l} is 0.77 for males and 0.70 for females.
}
}

\subsection{Lean body mass}{
\itemize{
\item "lbmb" : Calculates lean body mass (LBM), according to \href{https://pubmed.ncbi.nlm.nih.gov/6496691/}{Boers}:
\deqn{\textrm{LBM} = k \cdot \textrm{Weight (kg)} + l \cdot \textrm{Height (cm)} - m } , where \eqn{k} is 0.407 for males and 0.252 for females, and \eqn{l} is 0.267 for males and 0.473 for females, and \eqn{m} is 19.2 for males and 48.3 for females.
\item "lbmj" : Calculates lean body mass (LBM), according to \href{https://onlinelibrary.wiley.com/doi/10.1111/j.1467-3010.1977.tb00966.x}{James}:
\deqn{\textrm{LBM} = k \cdot \textrm{Weight (kg)} - l \cdot \left(\frac{\textrm{Weight (kg)}}{\textrm{Height (cm)}}\right)^2}, where \eqn{k} is 1.10 for males and 1.07 for females,  and  \eqn{l} is 128 in males and 148 in females.
\item "lbmp" : Calculates lean body mass (LBM) for children up to 14 years, according to \href{https://pubmed.ncbi.nlm.nih.gov/21498495/}{Peters}:
\deqn{\textrm{LBM} = 3.8 \cdot 0.0215 \cdot \textrm{Weight (kg)}^{0.6469} \cdot \textrm{Height (cm)}^{0.7236}}
}
}

\subsection{Predicted Normal Weight}{
\itemize{
\item "pnw" : Calculates the Predicted Normal Weight for obese patient, according to \href{https://pubmed.ncbi.nlm.nih.gov/15568893/}{Duffull}:
\deqn{\textrm{PNWT} = k \cdot  \textrm{Weight (kg)} - l \cdot \textrm{Height (cm)} \cdot \textrm{BMI} - m}, where \eqn{k} is 1.57 for males and 1.75 for females, and \eqn{l} is 0.0183 for males and 0.0242 for females, and \eqn{m} is 10.5 for males and 12.6 for females.
}
}
}
\examples{

tmp <- data.frame(id=1,WT=runif(3,70,120),HT=runif(3,160,220))
weight_height(wt=tmp$WT,ht=tmp$HT,type="bmi")
# example for use in dplyr
# tmp \%>\% mutate(BMI = weight_height(wt=WT,ht=HT,type="bmi"))
}
\author{
Richard Hooijmaijers
}
\keyword{misc}
