% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp.dm-package.R
\docType{package}
\name{amp.dm-package}
\alias{amp.dm}
\alias{amp.dm-package}
\title{amp.dm: data management tools for pharmacometrics}
\description{
Tools and functions to efficiently create datasets used in pharmacometric analysis. Additional functionality is added to create documention and prepare files for submission and QC purposes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LeidenAdvancedPKPD/amp.dm}
  \item Report bugs at \url{https://github.com/LeidenAdvancedPKPD/amp.dm/issues}
}

}
\author{
\strong{Maintainer}: Richard Hooijmaijers \email{richardhooijmaijers@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item LAPP Consultants \email{info@lapp.nl} [funder, copyright holder]
}

}
\keyword{internal}
