% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_df.r
\name{counts_df}
\alias{counts_df}
\title{Create counts and frequencies within in data frame}
\usage{
counts_df(
  data,
  by,
  id = NULL,
  style = 1,
  ret = "tbl",
  capt = "Information multiple data frames",
  align = NULL,
  ...
)
}
\arguments{
\item{data}{data frame for which the table should be created}

\item{by}{<\code{\link[rlang:args_data_masking]{data-masking}}> identifies by variables within the data frame for stratification}

\item{id}{<\code{\link[rlang:args_data_masking]{data-masking}}> identifies the ID variable within the data frame (see details)}

\item{style}{numeric with the type of output to return  (see details)}

\item{ret}{a character vector to define what kind of output should be returned (either "dfrm", "tbl", "file")}

\item{capt}{character with the caption of the table (not used in case data frame is returned)}

\item{align}{alignment of the table passed to \link{general_tbl} (not used in case data frame is returned)}

\item{...}{additional arguments passed to \link{general_tbl}}
}
\value{
a data frame, code for table or nothing in case a PDF file is created
}
\description{
This function calculates and reports counts and frequencies stratified by one or more variables within a
data frame
}
\details{
This function generates frequency tables, by default for the number of observation per strata. In case the id argument is used
the function will also report the number and frequencies of distinct IDs. By default the observations and percentages are reported in
separate columns (convenient for further processing). In case style is set to a value of 2, a single column is created that holds the
observations and percentages in a formatted ways (convenient for tabulating)
}
\examples{

data("Theoph")
Theoph$trt <- ifelse(as.numeric(Theoph$Subject)<6,1,2)
Theoph$sex <- ifelse(as.numeric(Theoph$Subject)<4,1,0)
counts_df(data=Theoph, by=c(trt,sex),id=Subject)
counts_df(data=Theoph, by=sex,id=Subject)
counts_df(data=Theoph, by=c(trt,sex),id=Subject, style=2)

}
\author{
Richard Hooijmaijers
}
\keyword{documentation}
