% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_log}
\alias{get_log}
\title{Retrieve log objects}
\usage{
get_log()
}
\value{
a named list of dataframes
}
\description{
Returns one or more dataframes with log information
related to function like \link{filterr}, \link{left_joinr}, \link{cmnt}, \link{srce} and \link{read_data}
}
\examples{
xldat <- readxl::readxl_example("datasets.xlsx")
xlin  <- read_data(xldat, comment="read test")
get_log()
}
\author{
Richard Hooijmaijers
}
\keyword{documentation}
